♻️ move Location CommandArgument into its own class
